# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
  /api/v1/register:
    post:
      summary: 'Register a new user'
      operationId: registerANewUser
      description: ''
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'User registered successfully'
                properties:
                  message:
                    type: string
                    example: 'User registered successfully'
      tags:
        - '1 Security'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: "The user's name"
                  example: eum
                  nullable: false
                email:
                  type: string
                  description: "The user's email"
                  example: koepp.shania@example.com
                  nullable: false
                password:
                  type: string
                  description: "The user's password"
                  example: 'tn^nC-pR".GRiuKr'
                  nullable: false
                phone:
                  type: string
                  description: "The user's phone number"
                  example: omnis
                  nullable: false
              required:
                - name
                - email
                - password
                - phone
      security: []
  /api/v1/login:
    post:
      summary: 'Login a user'
      operationId: loginAUser
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  token: string
                properties:
                  token:
                    type: string
                    example: string
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Invalid credentials'
                properties:
                  message:
                    type: string
                    example: 'Invalid credentials'
      tags:
        - '1 Security'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: "The user's email"
                  example: catherine.beatty@example.com
                  nullable: false
                password:
                  type: string
                  description: "The user's password"
                  example: 'gv78hVk!/_]{zq'
                  nullable: false
              required:
                - email
                - password
      security: []
  /api/v1/loginphone:
    post:
      summary: ''
      operationId: postApiV1Loginphone
      description: ''
      parameters: []
      responses: {  }
      tags:
        - '1 Security'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  description: 'Must not be greater than 15 characters.'
                  example: bxaxhpzr
                  nullable: false
              required:
                - phone
      security: []
...
  /api/v1/products:
    get:
      summary: 'Get a paginated list of products'
      operationId: getAPaginatedListOfProducts
      description: "No auth required.\nIf the quantity of a product is 0 or less, you cannot buy or add it to the cart.\nPagination: add ?page parameter to the request."
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "[\n   {\n     \"name\": \"Product 1\",\n     \"description\": \"Description of Product 1\",\n     \"price\": \"10.00 $\",\n     \"quantity\": 100\n   },\n   {\n     \"name\": \"Product 2\",\n     \"description\": \"Description of Product 2\",\n     \"price\": \"20.00 $\",\n     \"quantity\": 50\n   }\n ],"
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'No products found'
                properties:
                  message:
                    type: string
                    example: 'No products found'
      tags:
        - '2 Products'
      security: []
  /api/v1/menu:
    get:
      summary: 'Get a list of all menus'
      operationId: getAListOfAllMenus
      description: 'No auth required'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  menus:
                    -
                      id: 1
                      name: 'Menu 1'
                      image: images/menu1.jpg
                      description: 'Description of Menu 1'
                    -
                      id: 2
                      name: 'Menu 2'
                      image: images/menu2.jpg
                      description: 'Description of Menu 2'
                  lastMenu:
                    id: 1
                    name: 'Menu 1'
                    image: images/menu1.jpg
                    description: 'Description of Menu 1'
                properties:
                  menus:
                    type: array
                    example:
                      -
                        id: 1
                        name: 'Menu 1'
                        image: images/menu1.jpg
                        description: 'Description of Menu 1'
                      -
                        id: 2
                        name: 'Menu 2'
                        image: images/menu2.jpg
                        description: 'Description of Menu 2'
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: 'Menu 1'
                        image:
                          type: string
                          example: images/menu1.jpg
                        description:
                          type: string
                          example: 'Description of Menu 1'
                  lastMenu:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: 'Menu 1'
                      image:
                        type: string
                        example: images/menu1.jpg
                      description:
                        type: string
                        example: 'Description of Menu 1'
      tags:
        - '3 Menus'
      security: []
    post:
      summary: 'Create a new menu'
      operationId: createANewMenu
      description: 'Admin auth required'
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Menu created successfully'
                  data:
                    id: 1
                    name: 'Menu 1'
                    image: images/menu1.jpg
                    description: 'Description of Menu 1'
                properties:
                  message:
                    type: string
                    example: 'Menu created successfully'
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: 'Menu 1'
                      image:
                        type: string
                        example: images/menu1.jpg
                      description:
                        type: string
                        example: 'Description of Menu 1'
      tags:
        - '3 Menus'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: "The menu's name"
                  example: id
                  nullable: false
                image:
                  type: image
                  description: "The menu's image"
                  example: magnam
                  nullable: false
                description:
                  type: string
                  description: "The menu's description"
                  example: 'Repellat illo rerum facere minima aut consequuntur iusto.'
                  nullable: false
              required:
                - name
                - image
                - description
      security: []
...
  /api/v1/activitylog:
    get:
      summary: 'Display a listing of the resource.'
      operationId: displayAListingOfTheResource
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 1
                      log: 'User logged in'
                      user_id: 1
                      created_at: '2024-11-19T16:38:04.000000Z'
                      updated_at: '2024-11-19T16:38:04.000000Z'
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 1
                        log: 'User logged in'
                        user_id: 1
                        created_at: '2024-11-19T16:38:04.000000Z'
                        updated_at: '2024-11-19T16:38:04.000000Z'
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        log:
                          type: string
                          example: 'User logged in'
                        user_id:
                          type: integer
                          example: 1
                        created_at:
                          type: string
                          example: '2024-11-19T16:38:04.000000Z'
                        updated_at:
                          type: string
                          example: '2024-11-19T16:38:04.000000Z'
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Internal Server Error'
                properties:
                  message:
                    type: string
                    example: 'Internal Server Error'
      tags:
        - '111 Activity Log'
      security: []
...
  /api/v1/order:
    get:
      summary: 'Display a listing of the resource.'
      operationId: displayAListingOfTheResource
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    post:
      summary: 'Store a newly created resource in storage.'
      operationId: storeANewlyCreatedResourceInStorage
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: integer
                  description: ''
                  example: 4
                  nullable: false
                quantity:
                  type: integer
                  description: ''
                  example: 6
                  nullable: false
                price:
                  type: number
                  description: ''
                  example: 282.1912
                  nullable: false
                phone:
                  type: string
                  description: ''
                  example: aliquam
                  nullable: false
                addressIndetail:
                  type: string
                  description: ''
                  example: facilis
                  nullable: false
              required:
                - product_id
                - quantity
                - price
                - phone
                - addressIndetail
      security: []
...
  '/api/v1/updproduct2/{id}':
    post:
      summary: ''
      operationId: postApiV1Updproduct2Id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the updproduct2.'
        example: dolores
        required: true
        schema:
          type: string
...